[refactoring] extract the AttrDict class to a separate module in vissl.config: this allows to remove the implicit circular dependency between vissl.utils.hydra_config and vissl.config (which used to depend on hydra_config for the AttrDict class)
