+++ vissl/engines/extract_features.py
@@ -2,11 +2,12 @@
 
 import logging
 
+from vissl.config import AttrDict
 from vissl.trainer import SelfSupervisionTrainer
 from vissl.utils.checkpoint import get_checkpoint_folder
 from vissl.utils.collect_env import collect_env_info
 from vissl.utils.env import get_machine_local_and_dist_rank, set_env_vars
-from vissl.utils.hydra_config import AttrDict, print_cfg
+from vissl.utils.hydra_config import print_cfg
 from vissl.utils.io import save_file
 from vissl.utils.logger import setup_logging, shutdown_logging
 from vissl.utils.misc import set_seeds, setup_multiprocessing_method
