+++ vissl/hooks/__init__.py
@@ -4,6 +4,7 @@ from enum import Enum, auto
 from typing import List
 
 from classy_vision.hooks.classy_hook import ClassyHook
+from vissl.config import AttrDict
 from vissl.hooks.deepclusterv2_hooks import ClusterMemoryHook, InitMemoryHook  # noqa
 from vissl.hooks.grad_clip_hooks import GradClipHook  # noqa
 from vissl.hooks.log_hooks import (  # noqa
@@ -31,7 +32,6 @@ from vissl.hooks.swav_momentum_hooks import (
     SwAVMomentumNormalizePrototypesHook,
 )
 from vissl.hooks.tensorboard_hook import SSLTensorboardHook  # noqa
-from vissl.utils.hydra_config import AttrDict
 from vissl.utils.tensorboard import get_tensorboard_hook, is_tensorboard_available
