+++ vissl/losses/bce_logits_multiple_output_single_target.py
@@ -6,7 +6,7 @@ import torch
 from classy_vision.generic.util import is_on_gpu
 from classy_vision.losses import ClassyLoss, register_loss
 from torch import nn
-from vissl.utils.hydra_config import AttrDict
+from vissl.config import AttrDict
 
 
 @register_loss("bce_logits_multiple_output_single_target")
