+++ vissl/losses/deepclusterv2_loss.py
@@ -14,7 +14,7 @@ from classy_vision.generic.distributed_util import (
 )
 from classy_vision.losses import ClassyLoss, register_loss
 from torch import nn
-from vissl.utils.hydra_config import AttrDict
+from vissl.config import AttrDict
 from vissl.utils.misc import get_indices_sparse
