+++ vissl/losses/nce_loss.py
@@ -15,7 +15,7 @@ from classy_vision.generic.distributed_util import (
 from classy_vision.generic.util import is_pos_int
 from classy_vision.losses import ClassyLoss, register_loss
 from torch import nn
-from vissl.utils.hydra_config import AttrDict
+from vissl.config import AttrDict
 
 
 @register_loss("nce_loss_with_memory")
