+++ vissl/losses/simclr_info_nce_loss.py
@@ -8,8 +8,8 @@ import torch
 from classy_vision.generic.distributed_util import get_cuda_device_index, get_rank
 from classy_vision.losses import ClassyLoss, register_loss
 from torch import nn
+from vissl.config import AttrDict
 from vissl.utils.distributed_gradients import gather_from_all
-from vissl.utils.hydra_config import AttrDict
 
 
 @register_loss("simclr_info_nce_loss")
