+++ vissl/meters/mean_ap_list_meter.py
@@ -5,8 +5,8 @@ from typing import List, Union
 import torch
 from classy_vision.generic.util import is_pos_int
 from classy_vision.meters import ClassyMeter, register_meter
+from vissl.config import AttrDict
 from vissl.meters.mean_ap_meter import MeanAPMeter
-from vissl.utils.hydra_config import AttrDict
 
 
 @register_meter("mean_ap_list_meter")
