+++ vissl/models/heads/siamese_concat_view.py
@@ -2,8 +2,8 @@
 
 import torch
 import torch.nn as nn
+from vissl.config import AttrDict
 from vissl.models.heads import register_model_head
-from vissl.utils.hydra_config import AttrDict
 
 
 @register_model_head("siamese_concat_view")
