+++ vissl/models/heads/swav_prototypes_head.py
@@ -3,8 +3,8 @@ from typing import List
 
 import torch
 import torch.nn as nn
+from vissl.config import AttrDict
 from vissl.models.heads import register_model_head
-from vissl.utils.hydra_config import AttrDict
 
 
 @register_model_head("swav_head")
