+++ vissl/models/trunks/alexnet_jigsaw.py
@@ -1,9 +1,9 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 
 import torch.nn as nn
+from vissl.config import AttrDict
 from vissl.models.model_helpers import Flatten, get_trunk_forward_outputs_module_list
 from vissl.models.trunks import register_model_trunk
-from vissl.utils.hydra_config import AttrDict
 
 
 @register_model_trunk("alexnet_jigsaw")
