+++ vissl/models/trunks/regnet.py
@@ -6,13 +6,13 @@ from typing import List, Tuple
 import torch
 import torch.nn as nn
 from classy_vision.models import RegNet as ClassyRegNet, build_model
+from vissl.config import AttrDict
 from vissl.models.model_helpers import (
     Flatten,
     get_trunk_forward_outputs,
     transform_model_input_data_type,
 )
 from vissl.models.trunks import register_model_trunk
-from vissl.utils.hydra_config import AttrDict
 
 
 @register_model_trunk("regnet")
