+++ vissl/models/trunks/resnext.py
@@ -8,6 +8,7 @@ import torch
 import torch.nn as nn
 import torchvision.models as models
 from torchvision.models.resnet import Bottleneck
+from vissl.config import AttrDict
 from vissl.models.model_helpers import (
     Flatten,
     _get_norm,
@@ -15,7 +16,6 @@ from vissl.models.model_helpers import (
     transform_model_input_data_type,
 )
 from vissl.models.trunks import register_model_trunk
-from vissl.utils.hydra_config import AttrDict
 
 
 # For more depths, add the block config here
