+++ vissl/utils/hydra_config.py
@@ -6,84 +6,7 @@ import sys
 from typing import Any, List
 
 from omegaconf import DictConfig, OmegaConf
-from vissl.config import check_cfg_version
-
-
-class AttrDict(dict):
-    """
-    Dictionary subclass whose entries can be accessed like attributes (as well as normally).
-    Credits: https://aiida.readthedocs.io/projects/aiida-core/en/latest/_modules/aiida/common/extendeddicts.html#AttributeDict  # noqa
-    """
-
-    def __init__(self, dictionary):
-        """
-        Recursively turn the `dict` and all its nested dictionaries into `AttrDict` instance.
-        """
-        super().__init__()
-
-        for key, value in dictionary.items():
-            if isinstance(value, dict):
-                self[key] = AttrDict(value)
-            else:
-                self[key] = value
-
-    def __getattr__(self, key):
-        """
-        Read a key as an attribute.
-
-        :raises AttributeError: if the attribute does not correspond to an existing key.
-        """
-        if key in self:
-            return self[key]
-        else:
-            raise AttributeError(
-                f"{self.__class__.__name__} object has no attribute {key}."
-            )
-
-    def __setattr__(self, key, value):
-        """
-        Set a key as an attribute.
-        """
-        self[key] = value
-
-    def __delattr__(self, key):
-        """
-        Delete a key as an attribute.
-
-        :raises AttributeError: if the attribute does not correspond to an existing key.
-        """
-        if key in self:
-            del self[key]
-        else:
-            raise AttributeError(
-                f"{self.__class__.__name__} object has no attribute {key}."
-            )
-
-    def __getstate__(self):
-        """
-        Needed for pickling this class.
-        """
-        return self.__dict__.copy()
-
-    def __setstate__(self, dictionary):
-        """
-        Needed for pickling this class.
-        """
-        self.__dict__.update(dictionary)
-
-    def __deepcopy__(self, memo=None):
-        """
-        Deep copy.
-        """
-        from copy import deepcopy
-
-        if memo is None:
-            memo = {}
-        retval = deepcopy(dict(self))
-        return self.__class__(retval)
-
-    def __dir__(self):
-        return self.keys()
+from vissl.config import AttrDict, check_cfg_version
 
 
 def convert_to_attrdict(cfg: DictConfig, cmdline_args: List[Any] = None):
